Fix #63
My understanding of #63:
1. City is submitted through input box. Validation function fires a redirect, thus executing route code for a second time.
2. Original entry point attempts to execute remaining route code (generating an error in the process). New route skips validation, renders page.

To fix, I removed the function that wrapped the validation code and added return statements. Now when the validation function fires a redirect, it exits the code (preventing the error). The route then loads again (correctly) because of the redirect. I think better flow control should be implemented here, but it worked in my brief tests.